Version 1.2.0.0
- Supports syncing CGM data from Dexcom G4 Receivers with Nightscout via
USB and Bluetooth.

Updates:
- Added Bluetooth connectivity for Dexcom G4 Receivers
- Added a Progress Ring to the main page to better show sync status
- Added more detail to status messages when syncing
- Made it more clear when a new device is being enrolled
- Added a button to cancel synchronizations currently in progress

Bug Fixes:
- On startup when reading the list of enrolled devices the app could
crash if the list was corrupted or otherwise not properly formatted
- Made the hardware search process more efficient and doesn't leave the
app consuming a lot of CPU cycles once the process is complete
- Changed the format of uploaded dates and times to conform to ISO 8691

Known Issues:
- Occasionally the app will crash when closing communication with the
Dexcom Receiver when connected via USB.
- When the sync interval is shorter than 5 minutes and synchronizing
with the Dexcom Receiver over Bluetooth, the device going out of range
can cause future syncs to fail. To fix the issue, restart the app. The
workaround is to lengthen the sync interval to at least 5 minutes.